+++ scripts/api/ProjectionsManagerController.ts
@@ -3,19 +3,22 @@ import {injectable, inject} from 'inversify';
 import Dictionary from "../Dictionary";
 import IProjectionRunner from "../projections/IProjectionRunner";
 import {interfaces, Controller, Post} from 'inversify-express-utils';
+import {ISubject} from "rx";
 
 @Controller('/api/projections')
 @injectable()
 class ProjectionsManagerController implements interfaces.Controller {
 
-    constructor(@inject("IProjectionRunnerHolder") private projectionsRunnerCollection: Dictionary<IProjectionRunner<any>>) {
+    constructor(@inject("IProjectionRunnerHolder") private projectionsRunners: Dictionary<IProjectionRunner<any>>,
+                @inject("ProjectionStatuses") private projectionStatuses: ISubject<void>) {
     }
 
     @Post('/stop')
     stop(request: express.Request, response: express.Response): void {
         try {
-            this.getProjectionRunner(request.body.name).stop();
-            this.writeResponse(response, request.body.name, "Stop");
+            this.projectionsRunners[request.body.payload.name].stop();
+            this.projectionStatuses.onNext(null);
+            response.status(204).end();
         } catch (e) {
             response.status(400).json({error: "Projection not found or is already stopped"});
         }
@@ -24,8 +27,9 @@ class ProjectionsManagerController implements interfaces.Controller {
     @Post('/pause')
     pause(request: express.Request, response: express.Response): void {
         try {
-            this.getProjectionRunner(request.body.name).pause();
-            this.writeResponse(response, request.body.name, "Pause");
+            this.projectionsRunners[request.body.payload.name].pause();
+            this.projectionStatuses.onNext(null);
+            response.status(204).end();
         }
         catch (e) {
             response.status(400).json({error: "Projection not found or is not started"});
@@ -35,21 +39,15 @@ class ProjectionsManagerController implements interfaces.Controller {
     @Post('/resume')
     resume(request: express.Request, response: express.Response): void {
         try {
-            this.getProjectionRunner(request.body.name).resume();
-            this.writeResponse(response, request.body.name, "Resume");
+            this.projectionsRunners[request.body.payload.name].resume();
+            this.projectionStatuses.onNext(null);
+            response.status(204).end();
         }
         catch (e) {
             response.status(400).json({error: "Projection not found or is not paused"});
         }
     }
 
-    private getProjectionRunner(name: string): IProjectionRunner<any> {
-        return this.projectionsRunnerCollection[name];
-    }
-
-    private writeResponse(response: express.Response, name: string, operation: string) {
-        response.json({name: name, operation: operation});
-    }
 }
 
 export default ProjectionsManagerController;