+++ scripts/registry/ProjectionRegistry.ts
@@ -10,6 +10,7 @@ import IObjectContainer from "../bootstrap/IObjectContainer";
 import * as _ from "lodash";
 import ITickScheduler from "../ticks/ITickScheduler";
 import Dictionary from "../Dictionary";
+import {IProjection} from "../projections/IProjection";
 
 @injectable()
 class ProjectionRegistry implements IProjectionRegistry {
@@ -44,13 +45,6 @@ class ProjectionRegistry implements IProjectionRegistry {
         return this;
     }
 
-    private getDefinitionFromConstructor<T>(constructor:interfaces.Newable<IProjectionDefinition<T>>, area:string, name:string):IProjectionDefinition<T> {
-        const key = `prettygoat:definitions:${area}:${name}`;
-        if (!this.container.contains(key))
-            this.container.set(key, constructor);
-        return this.container.get<IProjectionDefinition<T>>(key);
-    }
-
     forArea(area:string):AreaRegistry {
         let entries = _.map(this.unregisteredEntries, entry => {
             let tickScheduler = <ITickScheduler>this.tickSchedulerFactory(),
@@ -64,7 +58,25 @@ class ProjectionRegistry implements IProjectionRegistry {
         let areaRegistry = new AreaRegistry(area, entries);
         this.registry.push(areaRegistry);
         this.unregisteredEntries = [];
+        if (this.hasDuplicatedEntries()) throw new Error("Cannot startup due to some projections with the same name");
         return areaRegistry;
+    }
+
+    private getDefinitionFromConstructor<T>(constructor:interfaces.Newable<IProjectionDefinition<T>>, area:string, name:string):IProjectionDefinition<T> {
+        const key = `prettygoat:definitions:${area}:${name}`;
+        if (!this.container.contains(key))
+            this.container.set(key, constructor);
+        return this.container.get<IProjectionDefinition<T>>(key);
+    }
+
+    private hasDuplicatedEntries():boolean {
+        let entries = _(this.getAreas())
+            .map((areaRegistry:AreaRegistry) => areaRegistry.entries)
+            .concat()
+            .flatten()
+            .groupBy((entry:RegistryEntry<any>) => entry.projection.name)
+            .valueOf();
+        return _.some(entries, (entry:RegistryEntry<any>[]) => entry.length > 1);
     }
 
     getAreas():AreaRegistry[] {